# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2.0"
  title: Laundree App
host: laundree.io
basePath: '/api'
schemes:
  - https
consumes:
  - application/json
  - application/octet-stream
produces:
  - application/json
securityDefinitions:
  auth0:
    authorizationUrl: "https://laundree-test.eu.auth0.com/authorize"
    flow: "implicit"
    type: "oauth2"
security:
  - jwt: []
paths:
...
  /laundries/{laundryId}:
    x-swagger-router-controller: laundries
    put:
      operationId: updateLaundry
      tags:
        - Laundries
      parameters:
        - name: laundryId
          in: path
          required: true
          type: string
        - name: updateLaundryBody
          in: body
          required: true
          schema:
            properties:
              name:
                type: string
              googlePlaceId:
                type: string
              rules:
                $ref: "#/definitions/Rules"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Laundry"
        "403":
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Laundry not found
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: Machine conflict
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"

    get:
      operationId: fetchLaundry
      tags:
        - Laundries
      parameters:
        - name: laundryId
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Laundry"
        "404":
          description: Laundry not found
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      operationId: deleteLaundry
      tags:
        - Laundries
      parameters:
        - name: laundryId
          in: path
          required: true
          type: string
      responses:
        "204":
          description: Success
        "404":
          description: Laundry not found
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Laundry:
    required:
      - id
      - name
      - href
      - owners
      - users
      - machines
      - invites
      - demo
      - timezone
      - googlePlaceId
      - rules
    properties:
      demo:
        type: boolean
      timezone:
        type: string
      googlePlaceId:
        type: string
      rules:
        $ref: "#/definitions/Rules"
      id:
        type: string
      name:
        type: string
      href:
        type: string
      owners:
        type: array
        items:
          $ref: "#/definitions/Summary"
      users:
        type: array
        items:
          $ref: "#/definitions/Summary"
      machines:
        type: array
        items:
          $ref: "#/definitions/Summary"
      invites:
        type: array
        items:
          $ref: "#/definitions/Summary"
  Rules:
    properties:
      limit:
        type: number
        format: int32
        minimum: 0
      dailyLimit:
        type: number
        format: int32
        minimum: 0
      timeLimit:
        required:
          - from
          - to
        properties:
          from:
            $ref: "#/definitions/Time"
          to:
            $ref: "#/definitions/Time"
...
  Summary:
    required:
      - id
      - href
    properties:
      id:
        type: string
      href:
        type: string
...
  Time:
    required:
      - hour
      - minute
    properties:
      hour:
        type: integer
      minute:
        type: integer
        enum:
          - 0
          - 30
  Error:
    required:
      - message
    properties:
      message:
        type: string
...